added idle state to player_state update

+++ includes/cub_utils.h
@@ -72,7 +72,7 @@ uint32_t	ft_strtoargb(const char *str, const char **str_ptr);
 void		*ft_read_all(const char *filename, size_t *file_size);
 t_vec2		vec2_rotate(t_vec2 vec, float angle);
 t_vec2		vec2_norm(t_vec2 vec);
-t_vec2		vec2_mult(t_vec2 a, int32_t scalar);
+t_vec2		vec2_mult(t_vec2 a, float scalar);
 float		vec2_dot(t_vec2 a, t_vec2 b);
 t_vec2		vec2_sum(t_vec2 a, t_vec2 b);
 int32_t		vec2_dist(t_vec2 a, t_vec2 b);
